error: save and load data not functioning correctly
