+++ GameFrameWork/GlobalClass/AbstractBoard.cs
