+++ GameFrameWork/GlobalClass/AbstractComputerPlayer.cs
