+++ GameFrameWork/GlobalClass/AbstractGame.cs
