+++ GameFrameWork/GlobalClass/AbstractHumanPlayer.cs
