+++ GameFrameWork/GlobalClass/AbstractPlayer.cs
