+++ GameFrameWork/GlobalClass/GameData.cs
