+++ GameFrameWork/GlobalClass/GameManager.cs
